\name{scholtzia}
\alias{scholtzia}
\docType{data}
\title{Australian bush plant Scholtzia}
\description{
  An inhomogeneous point pattern giving the locations of 171
  individuals of the bush \emph{Scholtzia} aff. \emph{involucrata} on a
  22 m x 22 m plot in Australian bush.
  
  Data were recorded and analysed by Armstrong (1991). A locally scaled point 
  process was fitted to the data in Prokesova et al. (2006). The data set was
  also analysed in Hahn and JEnsen (2013).
} 
\format{
  An object of class \code{"ppp"}
  representing the point pattern of individual plant locations.
  See \code{\link{ppp.object}} for details of the format.
}
\usage{data(scholtzia)}
\source{
  Armstrong (1991).
  By kind permission of Paul Armstrong.
}
\references{
  Armstrong, P. (1991). 
  Species patterning in heath vegetation of the northern sandplain. 
  Honours thesis, Murdoch University, Western Australia.
 
  Prokesova, M., Hahn, U., and Jensen, E.B.V. (2006). 
  Statistics for locally scaled point patterns. 
  In A. Baddeley, P. Gregori, J. Mateu, R. Stoica, and D. Stoyan (Eds.), 
  \emph{Case Studies in Spatial Point Pattern Modelling}, 
  number 185 in Lecture Notes in Statistics (pp. 99-123). New York: Springer-Verlag.
 
  Hahn, U. and Jensen, E. B. V. (2013)
  Inhomogeneous spatial point processes with hidden second-order stationarity.
  
} 
\keyword{datasets}
\keyword{spatial}

